W-9081821: Makes synchronous checkout the default checkout.
A flag in the convert examples file will now swap it to use the
asynchronous checkout.
The checkout-v2 was renamed checkout-async.
The framework checkout was removed entirely.
Updated readmes.
One script was almost entirely in an else condition. I just exit in the
if, removed the else and reduced tabbing of all lines by 1.
Still one outstanding issue is that I can't figure out how to set the
apex class to automatically be accessible by the b2b user.