W-8581201: Adds in support for promotions.
Renames the synchronous and asynchronous checkouts since that name is no
longer fully relevant.
Adds support for standard pricing to the Main flow (Synchronous flow).
By default this will not use standard pricing or promotions in the
main flow.
Adds a new set of navigation buttons that are required for the Main flow
to be able to go back to the previous screen when the standard pricing
or promotions are used.
Created several new subflows to support pricing and promotions.
Updated relevant scripts to have them support the new behavior.

Update 1: Adds a second main flow.
Several updates to the readmes associated with this change.
Fixed an installation issue with the renamed synchronous flows.
Update 2: Improves documentation per suggestion.
Update 3: Making the sync flow with promotions active. Fix typo.