W-9302034 and W-10275269: Adds cart to order sample apex back in.
Also fixes anchor navigation tabs on the flow to be buttons (better
matches the flow which uses buttons so as not to confuse users). This
also fixes some color formatting issues.

Additionally, the async subflows now pass a cart id on error to error
screen so as to avoid that ugly "No cart id" error we were seeing.